Fix du merge de hugo

+++ app/src/main/java/com/example/myapplication/slotmachine/SlotMachine.kt
@@ -4,7 +4,7 @@ import kotlin.math.atan
 import kotlin.math.max
 import kotlin.random.Random
 
-class SlotMachine (var luck: Int) {
+class SlotMachine (var luck: Float) {
     var fruits: Array<Fruit> = arrayOf()
         get() = field
         set(value){
